fix: documentation overhaul + readability and complexity improvements
