+++ src/components/Profile/ProfileMenu.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useAuth0 } from '@auth0/auth0-react';
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
-import Button from './Button';
+import Button from '../Button';
 import PropTypes from 'prop-types';
 
 /**
