+++ src/services/bestRouteService.js
@@ -14,7 +14,7 @@ const client = axios.create({
  * @param {number} start node ID for the start of the route
  * @param {number} end node ID for the end of the route
  * @param {string} skiArea ID of the ski area
- * @returns {Promise<{bestRoute: Object}>} bestRoute
+ * @returns {Promise<BestRoute>} bestRoute
  */
 export const fetchBestRoute = async (start, end, skiArea) => {
   try {
